pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "telco-call-centre"
version = "0.1.0"
description = "ML-based call centre ticket classification system"
authors = [{name = "Call Centre Team"}]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core ML & Data Science
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "datasets>=2.10.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "jupyter>=1.0.0",
    
    # FastAPI & Web Framework
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "pydantic[email]>=2.0.0",
    "python-multipart>=0.0.6",
    
    # Vector DB & RAG (Production-Ready)
    "pinecone-client>=3.0.0",
    "openai>=1.3.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-huggingface>=0.0.1",
    "chromadb>=0.4.20",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
    "weaviate-client>=3.25.0",
    "qdrant-client>=1.7.0",
    
    # LLM & AI (Offline-Capable)
    "ollama>=0.1.7",
    "huggingface-hub>=0.19.0",
    "accelerate>=0.24.0",
    "bitsandbytes>=0.41.0",
    "tokenizers>=0.15.0",
    
    # MLOps & Experiment Tracking
    "mlflow>=2.8.0",
    "prefect>=2.14.0",
    "dvc[s3]>=3.30.0",
    "wandb>=0.16.0",
    
    # Data Engineering & Pipelines
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.12.0",
    "redis>=5.0.0",
    "celery>=5.3.0",
    
    # NLP & Text Processing
    "spacy>=3.7.0",
    "nltk>=3.8.0",
    "textstat>=0.7.0",
    "python-dotenv>=1.0.0",
    
    # Data Validation & Quality
    "great-expectations>=0.18.0",
    "pydantic-settings>=2.0.0",
    "marshmallow>=3.20.0",
    
    # Enterprise Integrations
    "requests>=2.31.0",
    "requests-oauthlib>=1.3.0",
    "aiohttp>=3.9.0",
    "httpx>=0.25.0",
    
    # Security & Auth
    "cryptography>=41.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.0",
    "PyJWT>=2.8.0",
    
    # Observability & Monitoring
    "prometheus-client>=0.19.0",
    "structlog>=23.2.0",
    "rich>=13.7.0",
    "loguru>=0.7.0",
    
    # Utils & Helpers
    "faker>=20.0.0",
    "click>=8.1.0",
    "typer>=0.9.0",
    "tqdm>=4.66.0",
    "tenacity>=8.2.0",
    "backoff>=2.2.0",
]

[project.optional-dependencies]
dev = [
    # Testing Framework
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.25.0",  # For FastAPI testing
    
    # Code Quality & Formatting
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "pre-commit>=3.0.0",
    
    # Development Tools
    "ipython>=8.17.0",
    "jupyterlab>=4.0.0",
    "notebook>=7.0.0",
    
    # Documentation
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "sphinx>=7.2.0",
    
    # Profiling & Performance
    "memory-profiler>=0.61.0",
    "py-spy>=0.3.14",
    "scalene>=1.5.26",
]

production = [
    # Production-only dependencies
    "gunicorn>=21.2.0",
    "sentry-sdk[fastapi]>=1.38.0",
    "newrelic>=9.2.0",
    
    # Production Vector DBs (when online)
    "pinecone-client>=2.2.0",
    "weaviate-client>=3.25.0",
    "qdrant-client>=1.6.0",
    
    # Production LLM APIs (when online)
    "openai>=1.3.0",
    "anthropic>=0.7.0",
    "google-cloud-aiplatform>=1.38.0",
    
    # Production Enterprise Integrations
    "pysnow>=0.7.17",
    "zenpy>=2.0.25",
    "jira>=3.5.0",
    
    # Production Monitoring
    "datadog>=0.49.0",
    "grafana-api>=1.0.3",
    "elasticsearch>=8.11.0",
]

offline = [
    # Offline-specific models and data
    # Listed here for documentation purposes
]

[tool.ruff]
# Main configuration now in .ruff.toml
target-version = "py313"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml:coverage.xml",
    "--cov-fail-under=80",
    "--strict-markers",
    "-v"
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
